Move WrapGrpcContextError to globalerror
