+++ pkg/storegateway/storegatewaypb/custom.go
@@ -9,7 +9,7 @@ import (
 	"google.golang.org/grpc/metadata"
 
 	"github.com/grafana/mimir/pkg/storegateway/storepb"
-	"github.com/grafana/mimir/pkg/util"
+	"github.com/grafana/mimir/pkg/util/globalerror"
 )
 
 // customStoreGatewayClient is a custom StoreGatewayClient which wraps well known gRPC errors into standard golang errors.
@@ -27,7 +27,7 @@ func NewCustomStoreGatewayClient(cc *grpc.ClientConn) StoreGatewayClient {
 func (c *customStoreGatewayClient) Series(ctx context.Context, in *storepb.SeriesRequest, opts ...grpc.CallOption) (StoreGateway_SeriesClient, error) {
 	client, err := c.wrapped.Series(ctx, in, opts...)
 	if err != nil {
-		return client, util.WrapGrpcContextError(err)
+		return client, globalerror.WrapGrpcContextError(err)
 	}
 
 	return newCustomSeriesClient(client), nil
@@ -36,13 +36,13 @@ func (c *customStoreGatewayClient) Series(ctx context.Context, in *storepb.Serie
 // LabelNames implements StoreGatewayClient.
 func (c *customStoreGatewayClient) LabelNames(ctx context.Context, in *storepb.LabelNamesRequest, opts ...grpc.CallOption) (*storepb.LabelNamesResponse, error) {
 	res, err := c.wrapped.LabelNames(ctx, in, opts...)
-	return res, util.WrapGrpcContextError(err)
+	return res, globalerror.WrapGrpcContextError(err)
 }
 
 // LabelValues implements StoreGatewayClient.
 func (c *customStoreGatewayClient) LabelValues(ctx context.Context, in *storepb.LabelValuesRequest, opts ...grpc.CallOption) (*storepb.LabelValuesResponse, error) {
 	res, err := c.wrapped.LabelValues(ctx, in, opts...)
-	return res, util.WrapGrpcContextError(err)
+	return res, globalerror.WrapGrpcContextError(err)
 }
 
 // customStoreGatewayClient is a custom StoreGateway_SeriesClient which wraps well known gRPC errors into standard golang errors.
@@ -61,7 +61,7 @@ func newCustomSeriesClient(client StoreGateway_SeriesClient) *customSeriesClient
 
 func (c *customSeriesClient) Recv() (*storepb.SeriesResponse, error) {
 	res, err := c.wrapped.Recv()
-	return res, util.WrapGrpcContextError(err)
+	return res, globalerror.WrapGrpcContextError(err)
 }
 
 // customClientStream is a custom grpc.ClientStream which wraps well known gRPC errors into standard golang errors.
@@ -72,7 +72,7 @@ type customClientStream struct {
 // Header implements grpc.ClientStream.
 func (c *customClientStream) Header() (metadata.MD, error) {
 	md, err := c.wrapped.Header()
-	return md, util.WrapGrpcContextError(err)
+	return md, globalerror.WrapGrpcContextError(err)
 }
 
 // Trailer implements grpc.ClientStream.
@@ -83,7 +83,7 @@ func (c *customClientStream) Trailer() metadata.MD {
 // CloseSend implements grpc.ClientStream.
 func (c *customClientStream) CloseSend() error {
 	//nolint:forbidigo // Here we're just wrapping CloseSend() so it's OK to call it.
-	return util.WrapGrpcContextError(c.wrapped.CloseSend())
+	return globalerror.WrapGrpcContextError(c.wrapped.CloseSend())
 }
 
 // Context implements grpc.ClientStream.
@@ -93,10 +93,10 @@ func (c *customClientStream) Context() context.Context {
 
 // SendMsg implements grpc.ClientStream.
 func (c *customClientStream) SendMsg(m any) error {
-	return util.WrapGrpcContextError(c.wrapped.SendMsg(m))
+	return globalerror.WrapGrpcContextError(c.wrapped.SendMsg(m))
 }
 
 // RecvMsg implements grpc.ClientStream.
 func (c *customClientStream) RecvMsg(m any) error {
-	return util.WrapGrpcContextError(c.wrapped.RecvMsg(m))
+	return globalerror.WrapGrpcContextError(c.wrapped.RecvMsg(m))
 }
