+++ pkg/util/globalerror/user_test.go
